requirements.txt
pymupdf
numpy
pillow
fonttools
fastapi[all]
easyocr
google-cloud-storage
google-cloud-firestore
google-auth
python-dotenv
tensorflow
CacheControl